succesful validation

+++ mapt_flywheel_fixbids.py
@@ -10,21 +10,19 @@ bids_folder = "/Volumes/MAPT/fMRI/BIDS_data/"
 
 allsubjects = os.listdir(bids_folder)
 allsubjects.sort()
-print("Will check these:")
-print(allsubjects)
-to_do = []
+# to_do = []
 
-for subject_code in allsubjects:
+# for subject_code in allsubjects:
 
-    subject_folder = "%s/%s" % (bids_folder,subject_code)
-    inner_folder = "%s/ses-KaplanMAPT" % subject_folder
-    anat_folder = "%s/anat" % inner_folder
-    t1file = "%s/sub-%s_ses-KaplanMAPT_t1w.nii.gz" % (anat_folder,subject_code)
-    print("Checking %s" % t1file )
-    if (not os.path.exists(t1file)):
-        to_do.append(subject_code)
+#     subject_folder = "%s/%s" % (bids_folder,subject_code)
+#     inner_folder = "%s/ses-pre" % subject_folder
+#     anat_folder = "%s/anat" % inner_folder
+#     t1file = "%s/sub-%s_ses-pre_t1w.nii.gz" % (anat_folder,subject_code)
+#     print("Checking %s" % t1file )
+#     if (not os.path.exists(t1file)):
+#         to_do.append(subject_code)
 
-to_do = [elem for elem in to_do if "sub-" in elem]
+to_do = [elem for elem in allsubjects if "sub-" in elem]
 
 exclude = []
 
@@ -33,41 +31,54 @@ to_do = [elem for elem in to_do if elem not in exclude]
 
 print("Subjects to do:")
 print(to_do)
-to_do = ["sub-03191"]
 
 for subject_code in to_do:
+    print("Working on subject %s" % subject_code)
     subject_folder = "%s/%s" % (bids_folder,subject_code)
-    inner_folder = "%s/ses-KaplanMAPT" % subject_folder
-    fmap_folder = "%s/fmap" % inner_folder
-
+    inner_folders = ["ses-pre","ses-post"]
     subject_code = subject_code[4:]
 
-    # Fix capitalization of anatomicals
-    src_file = "%s/anat/sub-%s_ses-KaplanMAPT_t1w.json" % (inner_folder,subject_code)
-    dest_file = "%s/anat/sub-%s_ses-KaplanMAPT_T1w.json" % (inner_folder,subject_code)
-    if (os.path.exists(src_file)):
-        shutil.move(src_file,dest_file)
+    for inner_folder in inner_folders:
 
-    src_file = "%s/anat/sub-%s_ses-KaplanMAPT_t1w.nii.gz" % (inner_folder,subject_code)
-    dest_file = "%s/anat/sub-%s_ses-KaplanMAPT_T1w.nii.gz" % (inner_folder,subject_code)
-    if (os.path.exists(src_file)):
-        shutil.move(src_file,dest_file)
+        inner_folder_path = "%s/%s" % (subject_folder,inner_folder)
+        if os.path.exists(inner_folder_path):
+            print(" Working on session %s " % inner_folder)
 
-    src_file = "%s/anat/sub-%s_ses-KaplanMAPT_t2w.json" % (inner_folder,subject_code)
-    dest_file = "%s/anat/sub-%s_ses-KaplanMAPT_T2w.json" % (inner_folder,subject_code)
-    if (os.path.exists(src_file)):
-        shutil.move(src_file,dest_file)
+            fmap_folder = "%s/%s/fmap" % (subject_folder,inner_folder)
+            
+            # Fix capitalization of anatomicals
+            src_file = "%s/%s/anat/sub-%s_%s_t1w.json" % (subject_folder,inner_folder,subject_code,inner_folder)
+            dest_file = "%s/%s/anat/sub-%s_%s_T1w.json" % (subject_folder,inner_folder,subject_code,inner_folder)
+            print("Checking for %s" % src_file)
+            if (os.path.exists(src_file)):
+                print("Renaming %s to %s" % (src_file,dest_file))
+                shutil.move(src_file,dest_file)
 
-    src_file = "%s/anat/sub-%s_ses-KaplanMAPT_t2w.nii.gz" % (inner_folder,subject_code)
-    dest_file = "%s/anat/sub-%s_ses-KaplanMAPT_T2w.nii.gz" % (inner_folder,subject_code)
-    if (os.path.exists(src_file)):
-        shutil.move(src_file,dest_file)
+            src_file = "%s/%s/anat/sub-%s_%s_t1w.nii.gz" % (subject_folder,inner_folder,subject_code,inner_folder)
+            dest_file = "%s/%s/anat/sub-%s_%s_T1w.nii.gz" % (subject_folder,inner_folder,subject_code,inner_folder)
+            if (os.path.exists(src_file)):
+                print("Renaming %s to %s" % (src_file,dest_file))
+                shutil.move(src_file,dest_file)
 
-    fieldmap_jsons = os.listdir(fmap_folder)
-    for jsonfile in fieldmap_jsons:
-        command = "sed -i '' 's/t1w/T1w/g' %s/%s" % (fmap_folder,jsonfile)
-        print(command)
-        call(command,shell=True)
-        command = "sed -i '' 's/t2w/T2w/g' %s/%s" % (fmap_folder,jsonfile)
-        print(command)
-        call(command,shell=True)
+            src_file = "%s/%s/anat/sub-%s_%s_t2w.json" % (subject_folder,inner_folder,subject_code,inner_folder)
+            dest_file = "%s/%s/anat/sub-%s_%s_T2w.json" % (subject_folder,inner_folder,subject_code,inner_folder)
+            if (os.path.exists(src_file)):
+                print("Renaming %s to %s" % (src_file,dest_file))
+                shutil.move(src_file,dest_file)
+
+            src_file = "%s/%s/anat/sub-%s_%s_t2w.nii.gz" % (subject_folder,inner_folder,subject_code,inner_folder)
+            dest_file = "%s/%s/anat/sub-%s_%s_T2w.nii.gz" % (subject_folder,inner_folder,subject_code,inner_folder)
+            if (os.path.exists(src_file)):
+                print("Renaming %s to %s" % (src_file,dest_file))
+                shutil.move(src_file,dest_file)
+
+            if (os.path.exists(fmap_folder)):
+                print("Working fmap folder %s" % fmap_folder)
+                fieldmap_jsons = os.listdir(fmap_folder)
+                for jsonfile in fieldmap_jsons:
+                    command = "sed -i '' 's/t1w/T1w/g' %s/%s" % (fmap_folder,jsonfile)
+                    print(command)
+                    call(command,shell=True)
+                    command = "sed -i '' 's/t2w/T2w/g' %s/%s" % (fmap_folder,jsonfile)
+                    print(command)
+                    call(command,shell=True)
